Refactoring run_tests and run_tests_cli commands
