+++ th_cli/commands/__init__.py
@@ -16,7 +16,6 @@
 from .available_tests import available_tests
 from .project import create_project, delete_project, list_projects, update_project
 from .run_tests import run_tests
-from .run_tests_cli import run_tests_cli
 from .test_run_execution_history import test_run_execution_history
 from .test_runner_status import test_runner_status
 from .versions import versions
@@ -27,7 +26,6 @@ __all__ = [
     "delete_project",
     "list_projects",
     "run_tests",
-    "run_tests_cli",
     "test_run_execution_history",
     "update_project",
     "test_runner_status",
