--- th_cli/commands/run_tests_cli.py
@@ -1,158 +1,0 @@
-#
-# Copyright (c) 2025 Project CHIP Authors
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-import asyncio
-import datetime
-import json
-from typing import Optional
-
-import click
-
-import th_cli.api_lib_autogen.models as m
-import th_cli.test_run.logging as test_logging
-from th_cli.api_lib_autogen.api_client import AsyncApis
-from th_cli.api_lib_autogen.exceptions import UnexpectedResponse
-from th_cli.async_cmd import async_cmd
-from th_cli.client import get_client
-from th_cli.exceptions import CLIError, handle_api_error
-from th_cli.test_run.websocket import TestRunSocket
-from th_cli.utils import (
-    build_test_selection,
-    convert_nested_to_dict,
-    merge_properties_to_config,
-    read_pics_config,
-    read_properties_file,
-)
-from th_cli.validation import validate_directory_path, validate_file_path, validate_test_ids
-
-
-@click.command(no_args_is_help=True)
-@click.option(
-    "--tests-list",
-    "-t",
-    required=True,
-    help="List of test cases to execute. For example: TC-ACE-1.1,TC_ACE_1_3",
-)
-@click.option(
-    "--title",
-    "-n",
-    default=lambda: str(datetime.datetime.now().strftime("%Y-%m-%d-%H:%M:%S")),
-    show_default="timestamp",
-    help="Name of the test run execution",
-)
-@click.option(
-    "--config",
-    "-c",
-    help="Property config file location. This "
-    "information is optional — if not provided, the default_config.properties "
-    "file will be used.",
-)
-@click.option(
-    "--pics-config-folder",
-    "-p",
-    help="Directory containing PICS XML configuration files. If not provided, no PICS will be used.",
-)
-@async_cmd
-async def run_tests_cli(title: str, config: str, tests_list: str, pics_config_folder: str = None) -> None:
-    """Simplified CLI execution of a test run from selected tests"""
-
-    # Validate inputs and convert each test separated by comma to a list
-    validated_test_ids = validate_test_ids(tests_list)
-
-    if config:
-        config_path = validate_file_path(config, must_exist=True)
-        config = str(config_path)
-
-    if pics_config_folder:
-        pics_path = validate_directory_path(pics_config_folder, must_exist=True)
-        pics_config_folder = str(pics_path)
-
-    try:
-        client = get_client()
-        async_apis = AsyncApis(client)
-        projects_api = async_apis.projects_api
-        test_collections_api = async_apis.test_collections_api
-
-        # Configure new log output for test.
-        log_path = test_logging.configure_logger_for_run(title=title)
-
-        # Get default config and convert to dict
-        default_config = await projects_api.default_config_api_v1_projects_default_config_get()
-        default_config_dict = convert_nested_to_dict(default_config)
-
-        # If config file is provided, read and parse it
-        if not config:
-            config = "default_config.properties"
-
-        config_data = read_properties_file(config)
-        click.echo(f"Read config from file: {config_data}")
-        cli_config_dict = merge_properties_to_config(config_data, default_config_dict)
-        click.echo(f"CLI Config for test run execution: {cli_config_dict}")
-
-        # Read PICS configuration if provided
-        pics = read_pics_config(pics_config_folder)
-
-        test_collections = await test_collections_api.read_test_collections_api_v1_test_collections_get()
-        selected_tests_dict = build_test_selection(test_collections, validated_test_ids)
-
-        click.echo(f"Selected tests: {json.dumps(selected_tests_dict, indent=2)}")
-        new_test_run = await __create_new_test_run_cli(
-            async_apis, selected_tests=selected_tests_dict, title=title, config=cli_config_dict, pics=pics
-        )
-        socket = TestRunSocket(new_test_run)
-        socket_task = asyncio.create_task(socket.connect_websocket())
-        new_test_run = await __start_test_run(async_apis, new_test_run)
-        socket.run = new_test_run
-        await socket_task
-        click.echo(f"Log output in: '{log_path}'")
-    except CLIError:
-        raise  # Re-raise CLI errors
-    except Exception as e:
-        raise CLIError(f"Unexpected error during test execution: {e}")
-    finally:
-        if client:
-            await client.aclose()
-
-
-async def __create_new_test_run_cli(
-    async_apis: AsyncApis, selected_tests: dict, title: str, config: Optional[dict] = None, pics: Optional[dict] = None
-) -> m.TestRunExecutionWithChildren:
-    click.echo(f"Creating new test run with title: {title}")
-
-    test_run_in = m.TestRunExecutionCreate(title=title)
-    json_body = m.BodyCreateTestRunExecutionCliApiV1TestRunExecutionsCliPost(
-        test_run_execution_in=test_run_in, selected_tests=selected_tests, config=config, pics=pics
-    )
-
-    try:
-        test_run_executions_api = async_apis.test_run_executions_api
-        return await test_run_executions_api.create_test_run_execution_cli_api_v1_test_run_executions_cli_post(
-            json_body
-        )
-    except UnexpectedResponse as e:
-        handle_api_error(e, "create test run execution")
-
-
-async def __start_test_run(
-    async_apis: AsyncApis, test_run: m.TestRunExecutionWithChildren
-) -> m.TestRunExecutionWithChildren:
-    click.echo(f"Starting Test run: Title: {test_run.title}, id: {test_run.id}")
-    try:
-        test_run_executions_api = async_apis.test_run_executions_api
-        return await test_run_executions_api.start_test_run_execution_api_v1_test_run_executions_id_start_post(
-            id=test_run.id
-        )
-    except UnexpectedResponse as e:
-        handle_api_error(e, "start test run")
